adjusted 2fa style, page transition, login logic
